# apps/api/openapi.yml
openapi: 3.0.3
info:
  title: Risedle API
  version: 1.1.0
servers:
  - description: Staging
    url: https://apiedge.risedle.exchange
  - description: Production
    url: https://api.risedle.exchange
  - description: Local
    url: http://localhost:3000
...
paths:
  /v1/{chainId}/flts:
    get:
      tags:
        - Fuse Leveraged Token
      operationId: GetFuseLeveragedTokensByChainId
      summary: |
        Get list of Fuse Leveraged Tokens on specific chain
      parameters:
        - in: path
          name: chainId
          schema:
            $ref: "#/components/schemas/ChainId"
          required: true
      responses:
        200:
          description: |
            List of Fuse Leveaged Token and the market information
          content:
            application/json:
              schema:
                type: object
                properties:
                  tokens:
                    type: array
                    items:
                      $ref: "#/components/schemas/FuseLeveragedTokenInfo"
        404:
          description: Chain ID not supported
        500:
          description: Internal server error
...
components:
  schemas:
    Address:
      description: 20 byte Ethereum address encoded as a hex with `0x` prefix
      type: string
      example: "0xFF970A61A04b1cA14834A43f5dE4533eBDDB5CC8"
...
    ChainId:
      description: Chain ID
      type: number
      example: 42161
    USD:
      description: Price in USD terms
      type: number
      format: float
      example: 12.9879
    Percentage:
      description: Percentage
      type: number
      format: float
      example: 99.88888888
    FuseLeveragedTokenBackingInfo:
      description: Basic information of Fuse Leveraged Token underlying assets
      type: object
      properties:
        name:
          description: Asset's name
          type: string
          example: "Wrapped ETH"
        symbol:
          description: Asset's symbol
          type: string
          example: "WETH"
        amount:
          description: Asset's amount
          type: number
          format: float
          example: 1.2
        decimals:
          description: Asset's decimals
          type: number
          example: 18
        change:
          description: Last 24h change
          type: number
          format: float
          example: 1.2
        changePercent:
          description: Last 24h change in percentage
          type: number
          format: float
          example: 1.2
    FuseLeveragedTokenInfo:
      description: Basic information of Fuse Leveraged Token
      type: object
      properties:
        name:
          description: Fuse Leveraged Token's name
          type: string
          example: "2X Long ETH Risedle"
        symbol:
          description: Fuse Leveraged Token's symbol
          type: string
          example: "ETHRISE"
        decimals:
          description: Fuse Leveraged Token's decimals
          type: number
          example: 18
        address:
          description: Fuse Leveraged Token's contract address
          $ref: "#/components/schemas/Address"
        priceUSD:
          description: Current price in USD
          $ref: "#/components/schemas/USD"
        dailyPriceChangeUSD:
          description: |
            Price change (USD) in the last 24 hours. It can be negative.
          $ref: "#/components/schemas/USD"
        dailyPriceChangePercentage:
          description: |
            Price change (%) in the last 24 hours. It can be negative.
          $ref: "#/components/schemas/Percentage"
        totalVolumeUSD:
          description: Total trade volume in USD
          $ref: "#/components/schemas/USD"
        dailyVolumeChangeUSD:
          description: |
            Volume change (USD) in the last 24 hours. It can be negative.
          $ref: "#/components/schemas/USD"
        dailyVolumeChangePercentage:
          description: |
            Volume change (%) in the last 24 hours. It can be negative.
          $ref: "#/components/schemas/Percentage"
        marketcapUSD:
          description: Total supply times current price
          $ref: "#/components/schemas/USD"
        maxMarketcapUSD:
          description: Max supply times current price
          $ref: "#/components/schemas/USD"
        collateral:
          description: Collateral of Fuse Leveraged Token
          $ref: "#/components/schemas/FuseLeveragedTokenBackingInfo"
        debt:
          description: Debt of Fuse Leveraged Token
          $ref: "#/components/schemas/FuseLeveragedTokenBackingInfo"
        totalCollateral:
          description: Fuse Leveraged Token's total collateral
          type: number
          example: 1.1
        totalDebt:
          description: Fuse Leveraged Token's total debt
          type: number
          example: 999
...